调整登陆后获取用户资料

+++ app/src/main/java/com/thn/lexi/goods/explore/GoodsCollectionBean.java
@@ -119,13 +119,13 @@ public class GoodsCollectionBean {
                 public int material_id;
                 public String material_name;
                 public double max_price;
-                public int max_sale_price;
+                public double max_sale_price;
                 public double min_price;
-                public int min_sale_price;
+                public double min_sale_price;
                 public String name;
                 public int published_at;
                 public double real_price;
-                public int real_sale_price;
+                public double real_sale_price;
                 public String rid;
                 public int second_category_id;
                 public int status;
